+++ src/MythicScribe.ts
@@ -2,7 +2,7 @@ import * as vscode from 'vscode';
 
 import * as SubscriptionHelper from './subscriptions/SubscriptionHelper';
 import { getFormatter } from './formatter/formatter';
-import { addCustomDataset } from './datasets/customDatasets';
+import { addCustomDataset, removeCustomDataset } from './datasets/customDatasets';
 import { ScribeMechanicHandler } from './datasets/ScribeMechanic';
 import { ScribeEnumHandler } from './datasets/ScribeEnum';
 import { doVersionSpecificMigrations } from './migration/migration';
@@ -30,6 +30,7 @@ export async function activate(context: vscode.ExtensionContext) {
 
         // Commands
         vscode.commands.registerCommand('MythicScribe.addCustomDataset', addCustomDataset),
+        vscode.commands.registerCommand('MythicScribe.removeCustomDataset', removeCustomDataset),
 
         // Formatter
         getFormatter()